ANDROID ACCELEROMETER SYSTEM: TWEAK (TESTED SUCCESSFULLY ON 3 DEVICES SO FAR…)

+++ src/chronotext/android/cinder/CinderDelegate.cpp
@@ -93,10 +93,10 @@ namespace chronotext
     {
         struct AxisSwap
         {
-            float negateX;
-            float negateY;
-            float xSrc;
-            float ySrc;
+            int negateX;
+            int negateY;
+            int xSrc;
+            int ySrc;
         };
         
         static const AxisSwap axisSwap[] =
